+++ jest.setup.ts
@@ -1,8 +1,12 @@
 import "regenerator-runtime/runtime";
 import "@testing-library/jest-dom/extend-expect";
 import { TextDecoder, TextEncoder } from "util";
+import React from "react";
 import TimeAgo from "javascript-time-ago";
 import en from "javascript-time-ago/locale/en.json";
+import fetch from "node-fetch";
+import { lightTheme } from "@deskpro/deskpro-ui";
+import { mockClient } from "./src/testing";
 
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 //@ts-ignore
@@ -13,4 +17,47 @@ global.TextEncoder = TextEncoder;
 //@ts-ignore
 global.TextDecoder = TextDecoder;
 
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+//@ts-ignore
+global.React = React;
+
 TimeAgo.addDefaultLocale(en);
+
+jest.mock("@deskpro/app-sdk", () => ({
+  ...jest.requireActual("@deskpro/app-sdk"),
+  useDeskproAppClient: () => ({ client: mockClient }),
+  useDeskproAppEvents: (
+    hooks: { [key: string]: (param: Record<string, unknown>) => void },
+    deps: [] = []
+  ) => {
+    const deskproAppEventsObj = {
+      type: "ticket",
+      settings: {
+        instance_url: "http://instance.url",
+        permanent_auth_token: "this_is_token",
+      },
+      data: {
+        ticket: { id: "215", subject: "Big ticket" },
+        app: {},
+        env: {},
+        currentAgent: {},
+      },
+    };
+    React.useEffect(() => {
+      !!hooks.onChange && hooks.onChange(deskproAppEventsObj);
+      !!hooks.onShow && hooks.onShow(deskproAppEventsObj);
+      !!hooks.onReady && hooks.onReady(deskproAppEventsObj);
+      !!hooks.onAdminSettingsChange && hooks.onAdminSettingsChange(deskproAppEventsObj.settings);
+      /* eslint-disable-next-line react-hooks/exhaustive-deps */
+    }, deps);
+  },
+  useInitialisedDeskproAppClient: (callback: (param: Record<string, unknown>) => void) => {
+    callback({
+      registerElement: () => {},
+      deregisterElement: () => {},
+      setTitle: () => {},
+    });
+  },
+  useDeskproAppTheme: () => ({ theme: lightTheme }),
+  proxyFetch: async () => fetch,
+}));
