+++ src/testing/index.ts
@@ -1,1 +1,2 @@
 export { render } from "./render";
+export { mockClient } from "./mockClient";
